refactor auction filenames
