+++ src/features/auctions/command.ts
@@ -1,8 +1,8 @@
 import { Command } from "../../shared/command/command";
-import { itemAuctionSubcommand } from "./item-auction-subcommand";
-import { spellAuctionSubcommand } from "./spell-auction-subcommand";
+import { itemSubcommand } from "./item-subcommand";
+import { spellSubcommand } from "./spell-subcommand";
 
 export const auctionCommand = new Command("auction", "Start a new auction.", [
-  spellAuctionSubcommand,
-  itemAuctionSubcommand,
+  spellSubcommand,
+  itemSubcommand,
 ]);
