+++ src/features/auctions/spell-subcommand.ts
@@ -4,9 +4,9 @@ import {
   CommandInteraction,
 } from "discord.js";
 import { bankerRoleId } from "../../config";
-import { SpellAuctionThreadBuilder } from "./spell-auction-thread-builder";
+import { SpellThreadBuilder } from "./spell-thread-builder";
 import { ForbiddenSpells } from "../../shared/forbidden-spells";
-import { AuctionCommand, AuctionOption } from "./auction-base-subcommand";
+import { BaseSubcommand, BaseSubcommandOption } from "./base-subcommand";
 
 enum SpellOption {
   Player = "player",
@@ -14,14 +14,14 @@ enum SpellOption {
   ClassRole = "class",
 }
 
-export const Option = { ...SpellOption, ...AuctionOption };
+export const Option = { ...SpellOption, ...BaseSubcommandOption };
 
-class SpellAuction extends AuctionCommand {
+class Spell extends BaseSubcommand {
   public async execute(interaction: CommandInteraction<CacheType>) {
     const auctionChannel = await this.authorize(interaction);
 
     // send message to notify role
-    const builder = new SpellAuctionThreadBuilder(this.name, interaction);
+    const builder = new SpellThreadBuilder(this.name, interaction);
     const message = await auctionChannel.send(
       builder.classRole.map((r) => String(r)).join(" ")
     );
@@ -97,7 +97,7 @@ class SpellAuction extends AuctionCommand {
   }
 }
 
-export const spellAuctionSubcommand = new SpellAuction(
+export const spellSubcommand = new Spell(
   "spell",
   "Creates a new Forbidden Spell auction thread."
 );
